# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentasi API Belajar",
    "description": "Ini adalah API untuk aplikasi Belajar",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "https://backend-production-f9e7.up.railway.app/api/v1"
    },
    {
      "url": "http://localhost:9000/api/v1"
    }
  ],
  "paths": {
...
    "/content": {
      "get": {
        "tags": ["Content"],
        "summary": "Return hasil hit api untuk semua data content",
        "description": "ini API untuk mengambil semua data content",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sukses menampilkan data content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "object",
                      "example": "get all data content succesfull"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Data content kosong di database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Failed"
                    },
                    "message": {
                      "type": "object",
                      "example": "Data content is empty"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/content/{contentId}": {
      "get": {
        "tags": ["Content"],
        "summary": "Return hasil hit api data content dengan id",
        "description": "ini API untuk mengambil data content dengan id",
        "parameters": [
          {
            "name": "contentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Content ID"
          }
        ],
        "security": [
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sukses menampilkan data content dengan id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "object",
                      "example": "Get data content by id succesfull "
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Data content dengan id tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Failed"
                    },
                    "message": {
                      "type": "object",
                      "example": "Data content with id is empty"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerToken": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}